ratestask_app/requirements.txt
bottle
psycopg2-binary
waitress
